vectorcode: inherit `doCheck` from chromadb
The tests need to import onnxruntime (a transitive of chromadb),
however it has issues running in the build sandbox on aarch64-linux.

It may work at runtime, so the package is not marked broken, however it
will break tests, so the check phase is disabled when building on
aarch64-linux.

Since the issue comes via chromadb, we inherit its `doCheck` attribute.