nixos-rebuild-ng: allow `withNix218` to be overriden
`withNix218` controls if `nixos-rebuild-ng` is built with Nix 2.18+, so
it can use some features depending on newer version of Nix, for example
using `nix copy` instead of `nix-copy-closure` for more efficient copies
when both `--build-host` and `--target-host`. This is computed by
looking at the Nix derivation at compile time.

It is unlikely but it could happen that some specific use case an user
wants to use the Nix <2.18 path without actually using an older version
of Nix. For those cases we are exposing the `withNix218` as a derivation
parameter, so it can be overriden.