repart-image: do not run fixupPhase...
These derivation output one or more disk image files and metadata in
JSON, none of which seem to be good targets for patchelf, stripping
or other things that typically happen in fixupPhase.

I noticed that specifically shebang patching took a long time when
building bigger images, roughly doubling build times at times.

This might be due to a traversal of an entire disk images, dozens of
gigabytes in size while looking for interpreters.

All tests in

    nix-build -A nixosTests.systemd-repart

passed, but no further testing was done yet.